+++ src/main/java/io/dangernoodle/grt/steps/CreateRepositoryLabels.java
@@ -11,6 +11,7 @@ import io.dangernoodle.grt.GithubClient;
 import io.dangernoodle.grt.Repository;
 import io.dangernoodle.grt.Repository.Settings.Color;
 import io.dangernoodle.grt.Workflow.Context;
+import io.dangernoodle.grt.Workflow.Status;
 import io.dangernoodle.grt.internal.GithubWorkflow;
 
 
@@ -22,9 +23,9 @@ public class CreateRepositoryLabels extends GithubWorkflow.Step
     }
 
     @Override
-    public void execute(Repository repository, Context context) throws IOException
+    public Status execute(Repository repository, Context context) throws IOException
     {
-        GHRepository ghRepo = context.get(GHRepository.class);
+        GHRepository ghRepo = context.getGHRepository();
 
         Map<String, Color> labels = repository.getSettings().getLabels();
         Set<GHLabel> existing = ghRepo.listLabels().asSet();
@@ -48,6 +49,8 @@ public class CreateRepositoryLabels extends GithubWorkflow.Step
                 logger.info("label [{} / {}] already exists", name, color);
             }
         }
+        
+        return Status.CONTINUE;
     }
 
     private GHLabel findLabel(String name, Set<GHLabel> existing)