+++ src/test/java/io/dangernoodle/grt/steps/CreateRepositoryLabelsTest.java
@@ -11,7 +11,6 @@ import java.util.HashSet;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.kohsuke.github.GHLabel;
-import org.kohsuke.github.GHRepository;
 import org.kohsuke.github.PagedIterable;
 import org.mockito.Mock;
 
@@ -35,7 +34,6 @@ public class CreateRepositoryLabelsTest extends AbstractGithubWorkflowStepTest
 
         when(mockGHLabel.getName()).thenReturn("name");
         when(mockGHRepository.listLabels()).thenReturn(mockPagedIterable);
-        when(mockContext.get(GHRepository.class)).thenReturn(mockGHRepository);
     }
 
     @Test
@@ -44,6 +42,7 @@ public class CreateRepositoryLabelsTest extends AbstractGithubWorkflowStepTest
         givenALabelToCreate();
         whenExecuteStep();
         thenLabelWasCreated();
+        thenStatusIsContinue();
     }
 
     @Test
@@ -53,6 +52,7 @@ public class CreateRepositoryLabelsTest extends AbstractGithubWorkflowStepTest
         givenALabelThatAlreadyExists();
         whenExecuteStep();
         thenLabelWasNotCreated();
+        thenStatusIsContinue();
     }
 
     @Test
@@ -63,6 +63,7 @@ public class CreateRepositoryLabelsTest extends AbstractGithubWorkflowStepTest
         whenExecuteStep();
         // not the best but works for now
         thenLabelWasNotCreated();
+        thenStatusIsContinue();
     }
 
     @Override