+++ src/test/java/io/dangernoodle/grt/utils/RepositoryMergerTest.java
@@ -1,5 +1,6 @@
 package io.dangernoodle.grt.utils;
 
+import static io.dangernoodle.RepositoryAsserts.verifyArchived;
 import static io.dangernoodle.RepositoryAsserts.verifyBranchProtectionDisabled;
 import static io.dangernoodle.RepositoryAsserts.verifyBranchProtectionEnabled;
 import static io.dangernoodle.RepositoryAsserts.verifyCollaborators;
@@ -246,6 +247,14 @@ public class RepositoryMergerTest
     {
         whenBuildRepositories();
         thenImplicitDefaultsAreCorrect();
+    }
+
+    @Test
+    public void testImplicitOverrides()
+    {
+        givenImplictsAreOverridden();
+        whenBuildRepositories();
+        thenImplicitsAreOverridden();
     }
 
     @Test
@@ -552,6 +561,16 @@ public class RepositoryMergerTest
         deBuilder.addWorkflow("default");
     }
 
+    private void givenImplictsAreOverridden()
+    {
+        ovBuilder.setArchived(true)
+                 .setIssues(false)
+                 .setMergeCommits(false)
+                 .setRebaseMerge(false)
+                 .setSquashMerge(false)
+                 .setWiki(false);
+    }
+
     private void givenInitializeDefault()
     {
         deBuilder.setInitialize(true);
@@ -763,11 +782,22 @@ public class RepositoryMergerTest
 
     private void thenImplicitDefaultsAreCorrect()
     {
+        verifyArchived(repository, false);
         verifyIssues(repository, true);
         verifyMergeCommits(repository, true);
         verifyRebaseMerge(repository, true);
         verifySquashMerge(repository, true);
         verifyWiki(repository, true);
+    }
+
+    private void thenImplicitsAreOverridden()
+    {
+        verifyArchived(repository, true);
+        verifyIssues(repository, false);
+        verifyMergeCommits(repository, false);
+        verifyRebaseMerge(repository, false);
+        verifySquashMerge(repository, false);
+        verifyWiki(repository, false);
     }
 
     private void thenInitializationIsFalse()